Avoid setting adj to wrong process.
When pid wrap-around happens, there may be a new thread/process
uses a pid which has the same value as a previous just died process.

Before the death recipient reports, ActivityManager may still update
adj to the dying process, then the adj will set to wrong process.
Another extreme case is latency among socket that gives the timing.
Sometimes the victim may be thread of zygote or system server.

To avoid the case, comparing the real uid and incoming uid to have
higher confidence that it sets to the right one.

Change-Id: I3837334e04327f8c78549865fe3dc2b6121dcde4